ein paar kleinere bugfixes
