+++ .flake8
@@ -1,2 +1,3 @@
 [flake8]
 ignore = E501
+ignore = E722
